mono is now working on linux, fixed issues with exporting pointers and functions getting tripped.

+++ Mono/PlumbusEngine/PlumbusComponent.cs
@@ -4,7 +4,7 @@ namespace PlumbusEngine
 {
     public class PlumbusComponent : PlumbusObject
     {
-        public PlumbusComponent(IntPtr nativeObject, IntPtr gameObject)
+        public PlumbusComponent(UInt64 nativeObject, UInt64 gameObject)
          : base(nativeObject)
         {
             m_GameObject = gameObject;
@@ -15,6 +15,6 @@ namespace PlumbusEngine
             return new GameObject(m_GameObject);
         }
 
-        private IntPtr m_GameObject;
+        private UInt64 m_GameObject;
     }
 }
